Fixed a bug in memmove() - size_t(unsigned int) underflow Reimplemented Heap sort using different algorithm. Modified test case for sort.
git-svn-id: svn://svn.code.sf.net/p/aceos/code@281 78f6c469-4e3a-0410-bca2-dfd793a85426

+++ src/lib/string/mem_utils.c
@@ -56,7 +56,7 @@ void* memcpy(void *dest, const void *src, size_t n)
 */
 void* memmove(void *dest, void *src, size_t n)
 {
-    size_t i;
+    long int i;
 
     if((char*)src == (char*)dest)
     {
